Update Corrupt Database & server data location articles
This is an update to the corrupt database article following user feedback that it did not cover windows

Article re-written to mention other platforms and give the specific commands.

Also some adjustments - like saying that downgrade from beta to earlier public is generally not supported.

Also switched mentioned of `sudo -i` to `sudo -s` which is needed to give root privileges on some NAS without switching to root environment.

Also noticed that the server data folder article had inconsistent path for windows when compared to the other platforms. This is fixed and all platforms now give the path above the various subfolders

Also took opportunity to bring Plugins-Duplicate,md in-linen with Plugins which i updated recently. I do not know the background or reason for having a duplicate but it is now in line with the one it is a mirror copy of - except for the title